+++ spec/rubocop/cop/style/collection_methods_spec.rb
@@ -17,14 +17,14 @@ RSpec.describe RuboCop::Cop::Style::CollectionMethods, :config do
   cop_config['PreferredMethods'].each do |method, preferred_method|
     it "registers an offense for #{method} with block" do
       inspect_source("[1, 2, 3].#{method} { |e| e + 1 }")
-      expect(cop.offenses.size).to eq(1)
+
       expect(cop.messages)
         .to eq(["Prefer `#{preferred_method}` over `#{method}`."])
     end
 
     it "registers an offense for #{method} with proc param" do
       inspect_source("[1, 2, 3].#{method}(&:test)")
-      expect(cop.offenses.size).to eq(1)
+
       expect(cop.messages)
         .to eq(["Prefer `#{preferred_method}` over `#{method}`."])
     end
@@ -42,8 +42,8 @@ RSpec.describe RuboCop::Cop::Style::CollectionMethods, :config do
     end
 
     it 'auto-corrects to preferred method' do
-      new_source = autocorrect_source('some.collect(&:test)')
-      expect(new_source).to eq('some.map(&:test)')
+      new_source = autocorrect_source("some.#{method}(&:test)")
+      expect(new_source).to eq("some.#{preferred_method}(&:test)")
     end
   end
 end